+++ src/test/java/samples/types/GMonthDayNodeType.java
@@ -31,7 +31,7 @@ public final class GMonthDayNodeType extends ComparableNodeType <GMonthDay> {
 
 	
 	@Override
-	public Function<String, GMonthDay> valueConstructor() {
+	public Function<String, GMonthDay> getDataTypeConstructor() {
 		return GMonthDay.CONSTRUCTOR;
 	}