+++ src/test/java/samples/types/IBANNodeType.java
@@ -32,7 +32,7 @@ public final class IBANNodeType extends CharacterNodeType <IBAN> {
 	
 	
 	@Override
-	public Function<String, IBAN> valueConstructor() {
+	public Function<String, IBAN> getDataTypeConstructor() {
 		return IBAN.CONSTRUCTOR;
 	}